Conflicts:
	shotgun-client/app/common/components/OrderRequest.js
	shotgun-client/app/customer/CustomerOrderDetail.js
	shotgun-client/app/customer/dao/CustomerDao.js
	shotgun-client/app/driver/DriverOrderDetail.js
	shotgun-client/app/driver/DriverOrderRequestDetail.js
	viewserver-shotgun/src/main/java/com/shotgun/viewserver/ShotgunViewServerMaster.java

+++ viewserver-shotgun/src/main/java/com/shotgun/viewserver/payments/PaymentController.java
@@ -117,7 +117,7 @@ public class PaymentController {
     }
 
     @ControllerAction(path = "createCharge", isSynchronous = false)
-    public void createCharge(@ActionParam(name = "totalPrice") int totalPrice,
+    public void createCharge(@ActionParam(name = "totalPrice") Double totalPrice,
                              @ActionParam(name = "chargePercentage") int chargePercentage,
                              @ActionParam(name = "paymentId") String paymentId,
                              @ActionParam(name = "customerId") String customerId,
